cast int to float for return

+++ IMCCalculator.py
@@ -17,7 +17,7 @@ def main():
     imc = calculer_imc(args.poids, args.taille)
     print(imc)
     if imc is not None:
-        sys.exit(int(imc))
+        sys.exit(float(imc))
     else:
         print("Erreur lors du calcul de l'IMC.")
         sys.exit(1)
